feat(templates): prevent SignalR from re-connecting on auth state change #10918 (#10925)
